can't be worse but i'm trying

+++ src/OzonEdu.MerchandiseService.Infrastructure/Commands/CreateMerchRequest/CreateMerchRequestHistoryEntryCommand.cs
@@ -6,7 +6,7 @@ namespace OzonEdu.MerchandiseService.Infrastructure.Commands.CreateMerchRequest
 {
     public class CreateMerchRequestHistoryEntryCommand : IRequest
     {
-        public long EmployeeId { get; init; }
+        public int EmployeeId { get; init; }
         
         public string MerchPackName { get; init; }
