+++ src/OzonEdu.MerchandiseService.Infrastructure/DomainServices/Interfaces/IMerchRequestDomainService.cs
@@ -7,10 +7,10 @@ namespace OzonEdu.MerchandiseService.Infrastructure.DomainServices.Interfaces
 {
     public interface IMerchRequestDomainService
     {
-        Task GiveOutMerch(long employeeId, string merchPackName, Dictionary<string, string> constraints,
+        Task GiveOutMerch(int employeeId, string merchPackName, Dictionary<string, string> constraints,
             CancellationToken token);
 
-        Task<List<MerchIssuedToEmployee>> GetMerchIssuedToEmployee(long employeeId,
+        Task<List<MerchIssuedToEmployee>> GetMerchIssuedToEmployee(int employeeId,
             CancellationToken token);
     }
 }
